Use consistent marginal implementation for all TracePosterior subclasses (#1649)
* Use consistent marginal implementation for all TracePosterior subclasses

* clarify documentation

* clarify documentation
